字段修改

+++ src/main/java/cn/com/bsfit/frms/policy/mapper/RiskArchivesMapper.java
@@ -12,9 +12,9 @@ public interface RiskArchivesMapper {
 
     int deleteByPrimaryKey(Long archivesId);
 
-    long insert(RiskArchives record);
+    int insert(RiskArchives record);
 
-    long insertSelective(RiskArchives record);
+    int insertSelective(RiskArchives record);
 
     List<RiskArchives> selectByExample(RiskArchivesExample example);
